Made request for data earlier
