+++ src/app/components/engangsstonad/EngangsstonadConfirmation.js
@@ -3,6 +3,7 @@ import { connect } from 'react-redux';
 import PropTypes from 'prop-types';
 import DocumentTitle from 'react-document-title';
 import { withRouter } from 'react-router-dom';
+import queryStringParser from 'query-string';
 
 import { Ingress } from 'nav-frontend-typografi';
 import { Hovedknapp } from 'nav-frontend-knapper';
@@ -10,7 +11,7 @@ import { Hovedknapp } from 'nav-frontend-knapper';
 import ConfirmCheckbox from 'shared/confirmCheckbox/ConfirmCheckbox';
 import HeaderIllustration from 'shared/header-illustration/HeaderIllustration';
 import VelkommenIllustration from 'assets/svg/frontpage.svg';
-import { approveConditions } from 'actions';
+import { getDataRequested, approveConditions } from 'actions';
 
 import './engangsstonad.less';
 
@@ -26,6 +27,20 @@ export class EngangsstonadConfirmation extends Component {
 		);
 	}
 
+	componentWillMount() {
+		const queryParams = this.getQueryParams();
+
+		if (Object.keys(queryParams).length > 0) {
+			this.props.dispatch(getDataRequested(queryParams));
+		} else {
+			this.props.dispatch(getDataRequested());
+		}
+	}
+
+	getQueryParams() {
+		return queryStringParser.parse(this.props.location.search);
+	}
+
 	handleConfirmCheckboxChange() {
 		this.props.dispatch(approveConditions());
 	}
@@ -35,13 +50,18 @@ export class EngangsstonadConfirmation extends Component {
 	}
 
 	render() {
-		const { approvedConditions } = this.props;
+		const { data, approvedConditions } = this.props;
+
+		if (!data) {
+			return null;
+		}
+
 		return (
 			<div className="engangsstonad">
 				<DocumentTitle title="Samtykke - NAV Engangsstønad" />
 				<HeaderIllustration
 					dialog={{
-						title: 'Hei, Test',
+						title: `Hei, ${this.props.data.fornavn}`,
 						text: 'Jeg skal hjelpe deg med å fylle ut søknaden'
 					}}
 					svg={VelkommenIllustration}
@@ -75,14 +95,22 @@ EngangsstonadConfirmation.propTypes = {
 	dispatch: PropTypes.func.isRequired,
 	history: PropTypes.shape({
 		push: PropTypes.func.isRequired
-	}).isRequired
+	}).isRequired,
+	location: PropTypes.shape({
+		search: PropTypes.string
+	}).isRequired,
+	data: PropTypes.shape({
+		fornavn: PropTypes.string
+	})
 };
 
 EngangsstonadConfirmation.defaultProps = {
-	approvedConditions: false
+	approvedConditions: false,
+	data: undefined
 };
 
 const mapStateToProps = (state) => ({
+	data: state.engangsstonadReducer.data,
 	approvedConditions: state.engangsstonadReducer.approvedConditions
 });
